requirements.txt
pygame-ce
numpy